Fix circular dependency, now launches up to tmux (testing on windows lol)
